[Library] Rename library name and enable to use from pip.
