+++ .pytest_cache/v/cache/lastfailed
@@ -1,0 +1,3 @@
+{
+  "nmf_nn/tests/test_np_biased_nmf.py": true
+}
