--- benchmark_model/__init__.py
@@ -1,1 +1,0 @@
-from .double_model import *
