--- matrix_factorization/__init__.py
@@ -1,2 +1,0 @@
-from matrix_factorization.matrix_factorization import semi_nmf, nonlin_semi_nmf
-from matrix_factorization.np_biased_nmf import _nonlin_solve
