--- optimizer/__init__.py
@@ -1,1 +1,0 @@
-from optimizer.optimizers import NMFOptimizer
