+++ sakurai_nmf/__init__.py
@@ -1,0 +1,12 @@
+"""Main package of NMF neural network."""
+
+from __future__ import absolute_import
+from __future__ import division
+from __future__ import print_function
+
+from . import benchmark_model
+from . import examples
+from . import matrix_factorization
+from . import optimizer
+from . import tests
+from . import losses
