+++ sakurai_nmf/benchmark_model/__init__.py
@@ -1,0 +1,2 @@
+from .double_model import *
+from . import dataset
