+++ sakurai_nmf/benchmark_model/mnist_dense.py
