+++ sakurai_nmf/benchmark_model/rnn_model.py
