+++ sakurai_nmf/examples/__init__.py
@@ -1,0 +1,2 @@
+from . import mnist
+from . import cifar10
