+++ sakurai_nmf/examples/mnist.py
@@ -1,38 +1,58 @@
 """Main of (fashion) mnist model."""
 
-import agents
+from __future__ import absolute_import
+from __future__ import division
+from __future__ import print_function
+
 import functools
 import numpy as np
 import tensorflow as tf
+from agents.tools import AttrDict
 
-import benchmark_model
-from optimizer import NMFOptimizer
+from sakurai import benchmark_model
+from sakurai.optimizer import NMFOptimizer
 
 
 def default_config():
     # Batch size
-    batch_size = benchmark_model.batch_size
+    batch_size = FLAGS.batch_size
+    # Dataset
+    dataset = FLAGS.dataset
     # Number of matrix factorization iterations
-    num_mf_iters = 0
+    num_mf_iters = FLAGS.num_mf_iters
     # Number of back propagation iterations
-    num_bp_iters = 15
+    num_bp_iters = FLAGS.num_bp_iters
     # Learning rate for adam
-    learning_rate = 0.01
+    learning_rate = FLAGS.lr
     # NMF actiovation
-    activation = None
+    if FLAGS.use_relu:
+        activation = tf.nn.relu
+    else:
+        activation = None
     # NMF use bias
-    use_bias = True
+    use_bias = FLAGS.use_bias
     return locals()
 
 
-def train_and_test(train_op, num_iters, sess, model, x_train, y_train, x_test, y_test, batch_size=1):
+def train_and_test(train_op, num_iters, sess, model, x_train, y_train, x_test, y_test, batch_size=1,
+                   output_debug=False):
     for i in range(num_iters):
         # Train...
         x, y = benchmark_model.batch(x_train, y_train, batch_size=batch_size)
-        _, train_loss, train_acc = sess.run([train_op, model.cross_entropy, model.accuracy], feed_dict={
+        _, = sess.run([train_op], feed_dict={
             model.inputs: x,
             model.labels: y,
         })
+        train_loss, train_acc = sess.run([model.cross_entropy, model.accuracy], feed_dict={
+            model.inputs: x,
+            model.labels: y,
+        })
+        if output_debug:
+            outputs = sess.run(model.outputs, feed_dict={
+                model.inputs: x,
+                model.labels: y,
+            })
+            print(outputs)
         stats = []
         # Compute test accuracy.
         for _ in range(5):
@@ -43,7 +63,7 @@ def train_and_test(train_op, num_iters, sess, model, x_train, y_train, x_test, y
             }))
         test_loss, test_acc = np.mean(stats, axis=0)
         
-        print('\r({}/{}) [Train]loss {}, accuracy {} [Test]loss {}, accuracy {}'.format(
+        print('\r({}/{}) [Train]loss {:.3f}, accuracy {:.3f} [Test]loss {:.3f}, accuracy {:.3f}'.format(
             i + 1, num_iters,
             train_loss, train_acc, test_loss, test_acc), end='', flush=True)
     print()
@@ -51,15 +71,17 @@ def train_and_test(train_op, num_iters, sess, model, x_train, y_train, x_test, y
 
 def main(_):
     # Set configuration
-    config = agents.tools.AttrDict(default_config())
+    config = AttrDict(default_config())
     # Build one hot mnist model.
-    model = benchmark_model.build_tf_one_hot_model(3072, use_bias=config.use_bias, activation=config.activation)
+    model = benchmark_model.build_tf_one_hot_model(batch_size=config.batch_size,
+                                                   use_bias=config.use_bias,
+                                                   activation=config.activation)
     # Load one hot mnist data.
-    (x_train, y_train), (x_test, y_test) = benchmark_model.load_one_hot_data(dataset='cifar10')
+    (x_train, y_train), (x_test, y_test) = benchmark_model.load_one_hot_data(dataset=config.dataset)
     
     # Testing whether the dataset have correct shape.
-    assert x_train.shape == (50000, 3072)
-    assert y_train.shape == (50000, 10)
+    assert x_train.shape == (60000, 784)
+    assert y_train.shape == (60000, 10)
     
     # Minimize model's loss with NMF optimizer.
     # optimizer = NMFOptimizer(config)
@@ -89,4 +111,13 @@ def main(_):
 
 
 if __name__ == '__main__':
+    FLAGS = tf.app.flags.FLAGS
+    tf.app.flags.DEFINE_integer('batch_size', 3000, """Size of batches""")
+    tf.app.flags.DEFINE_string('dataset', 'mnist', '''mnist or fashion''')
+    tf.app.flags.DEFINE_integer('num_mf_iters', 3, '''Number of matrix factorization iterations''')
+    tf.app.flags.DEFINE_integer('num_bp_iters', 5, '''Number of back propagation(adam) iterations''')
+    tf.app.flags.DEFINE_float('lr', 0.001, '''learning rate for back propagation''')
+    tf.app.flags.DEFINE_boolean('use_relu', False, '''Use ReLU''')
+    tf.app.flags.DEFINE_boolean('use_bias', True, '''Use bias''')
+    
     tf.app.run()
