+++ sakurai_nmf/exception.py
