+++ sakurai_nmf/losses.py
