+++ sakurai_nmf/matrix_factorization/__init__.py
@@ -1,0 +1,1 @@
+from .matrix_factorization import *
