+++ sakurai_nmf/matrix_factorization/utility.py
