+++ sakurai_nmf/optimizer/__init__.py
@@ -1,0 +1,1 @@
+from .optimizers import NMFOptimizer
