+++ sakurai_nmf/optimizer/assets/result_opt.png
