+++ sakurai_nmf/optimizer/utility.py
