+++ sakurai_nmf/tensorflow_test/np_svd.py
