+++ sakurai_nmf/tensorflow_test/vanilla_tf_svd.py
