+++ sakurai_nmf/tests/generate_test_mat.py
