+++ sakurai_nmf/tests/test_detail_function.py
