+++ sakurai_nmf/tests/test_matrix_factorization.py
