+++ sakurai_nmf/tests/test_np_biased_nmf.py
