+++ sakurai_nmf/tests/test_np_nmf.py
