+++ sakurai_nmf/tests/test_optimizers.py
