+++ sakurai_nmf/tests/test_tf_matrix_factorization.py
