+++ sakurai_nmf/tests/test_utility.py
