+++ sakurai_nmf/utils.py
